requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
black==26.3.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.3
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==8.0.1
joblib==1.5.3
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.20.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
